Changed template to use doubles

+++ src/MainTemplate.java
@@ -123,14 +123,14 @@ abstract class DrawableObject {
 	private Consumer<DrawableObject> onClicked = input -> {
 	}; // Default empty lambda
 
-	public DrawableObject(int x, int y, Color color) {
+	public DrawableObject(double x, double y, Color color) {
 		this.x = x;
 		this.y = y;
 		this.color = color;
 	}
 
 	// Moves the object to the specified position
-	public void moveTo(int x, int y) {
+	public void moveTo(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -154,7 +154,7 @@ abstract class DrawableObject {
 	public abstract void draw(Graphics g);
 
 	// Used for checking if a point is inside the object (for mouse clicks)
-	public abstract boolean contains(int x, int y);
+	public abstract boolean contains(double x, double y);
 }
 
 class Box extends DrawableObject {
@@ -166,13 +166,13 @@ class Box extends DrawableObject {
 		this.height = 100;
 	}
 
-	public Box(int x, int y, int width, int height, Color color) {
+	public Box(double x, double y, double width, double height, Color color) {
 		super(x, y, color);
 		this.width = width;
 		this.height = height;
 	}
 
-	public void setSize(int width, int height) {
+	public void setSize(double width, double height) {
 		this.width = width;
 		this.height = height;
 	}
@@ -189,7 +189,7 @@ class Box extends DrawableObject {
 	}
 
 	@Override
-	public boolean contains(int x, int y) {
+	public boolean contains(double x, double y) {
 		return x >= this.x && x <= this.x + width && y >= this.y && y <= this.y + height;
 	}
 }
@@ -223,7 +223,7 @@ class Circle extends DrawableObject {
 	}
 
 	@Override
-	public boolean contains(int x, int y) {
+	public boolean contains(double x, double y) {
 		double dx = this.x - x;
 		double dy = this.y - y;
 		return dx * dx + dy * dy <= radius * radius;
@@ -275,7 +275,7 @@ class Line extends DrawableObject {
 	}
 
 	@Override
-	public boolean contains(int x, int y) {
+	public boolean contains(double x, double y) {
 		return false; // Lines are not clickable (yet?)
 	}
 }
@@ -289,7 +289,7 @@ class Text extends DrawableObject {
 		this.text = "";
 	}
 
-	public Text(int x, int y, String text, Color color) {
+	public Text(double x, double y, String text, Color color) {
 		super(x, y, color);
 		this.text = text;
 	}
@@ -322,7 +322,7 @@ class Text extends DrawableObject {
 	}
 
 	@Override
-	public boolean contains(int x, int y) {
+	public boolean contains(double x, double y) {
 		// Text objects are not clickable, so always return false
 		return false;
 	}
